+++ src/Modules/Quiz/QuizXhr.php
@@ -6,26 +6,28 @@ use Foodsharing\Helpers\DataHelper;
 use Foodsharing\Lib\Xhr\XhrDialog;
 use Foodsharing\Modules\Content\ContentGateway;
 use Foodsharing\Modules\Core\Control;
+use Foodsharing\Modules\Core\DBConstants\Foodsaver\Role;
 use Foodsharing\Services\SanitizerService;
 
 class QuizXhr extends Control
 {
 	private $contentGateway;
 	private $quizGateway;
+	private $quizSessionGateway;
 	private $sanitizerService;
 	private $dataHelper;
 
 	public function __construct(
-		QuizModel $model,
 		QuizGateway $quizGateway,
+		QuizGateway $quizSessionGateway,
 		QuizView $view,
 		ContentGateway $contentGateway,
 		SanitizerService $sanitizerService,
 		DataHelper $dataHelper
 	) {
-		$this->model = $model;
 		$this->view = $view;
 		$this->quizGateway = $quizGateway;
+		$this->quizSessionGateway = $quizSessionGateway;
 		$this->contentGateway = $contentGateway;
 		$this->sanitizerService = $sanitizerService;
 		$this->dataHelper = $dataHelper;
@@ -35,7 +37,7 @@ class QuizXhr extends Control
 
 	public function hideinfo()
 	{
-		$this->session->setOption('quiz-infobox-seen', true, $this->model);
+		$this->session->setOption('quiz-infobox-seen', true);
 	}
 
 	public function addquest()
@@ -49,12 +51,13 @@ class QuizXhr extends Control
 		 */
 		if ($this->session->mayEditQuiz()) {
 			if (isset($_GET['text'], $_GET['fp'], $_GET['qid'])) {
-				$fp = (int)$_GET['fp'];
+				$failurePoints = (int)$_GET['fp'];
 				$text = strip_tags($_GET['text']);
 				$duration = (int)$_GET['duration'];
 
 				if (!empty($text)) {
-					if ($id = $this->model->addQuestion($_GET['qid'], $text, $fp, $duration)) {
+					$id = $this->quizGateway->addQuestion($_GET['qid'], $text, $failurePoints, $duration);
+					if ($id > 0) {
 						$this->flashMessageHelper->info('Frage wurde angelegt');
 
 						return array(
@@ -75,7 +78,7 @@ class QuizXhr extends Control
 	public function delquest()
 	{
 		if ($this->session->mayEditQuiz() && isset($_GET['id'])) {
-			$this->model->deleteQuest($_GET['id']);
+			$this->quizGateway->deleteQuestion($_GET['id']);
 
 			return array(
 				'status' => 1,
@@ -87,7 +90,7 @@ class QuizXhr extends Control
 	public function delanswer()
 	{
 		if ($this->session->mayEditQuiz() && isset($_GET['id'])) {
-			$this->model->deleteAnswer($_GET['id']);
+			$this->quizGateway->deleteAnswer($_GET['id']);
 
 			return array(
 				'status' => 1,
@@ -112,7 +115,7 @@ class QuizXhr extends Control
 				$right = (int)$_GET['right'];
 
 				if (!empty($text) && ($right == 0 || $right == 1 || $right == 2)) {
-					if ($id = $this->model->addAnswer($_GET['qid'], $text, $exp, $right)) {
+					if ($id = $this->quizGateway->addAnswer($_GET['qid'], $text, $exp, $right)) {
 						return array(
 							'status' => 1,
 							'script' => 'pulseInfo("Antwort wurde angelegt");$("#answerlist-' . (int)$_GET['qid'] . '").append(\'<li class="right-' . (int)$right . '">' . $this->sanitizerService->jsSafe(nl2br(strip_tags($text))) . '</li>\');$( "#questions" ).accordion( "refresh" );'
@@ -137,11 +140,7 @@ class QuizXhr extends Control
 				$right = (int)$_GET['right'];
 
 				if (!empty($text) && ($right == 0 || $right == 1 || $right == 2)) {
-					$this->model->updateAnswer($_GET['id'], array(
-						'text' => $text,
-						'explanation' => $exp,
-						'right' => $right
-					));
+					$this->quizGateway->updateAnswer($_GET['id'], $text, $exp, $right);
 
 					return array(
 						'status' => 1,
@@ -160,7 +159,7 @@ class QuizXhr extends Control
 	public function editanswer()
 	{
 		if ($this->session->mayEditQuiz()) {
-			if ($answer = $this->model->getAnswer($_GET['id'])) {
+			if ($answer = $this->quizGateway->getAnswer($_GET['id'])) {
 				$this->dataHelper->setEditData($answer);
 				$dia = new XhrDialog();
 
@@ -237,7 +236,7 @@ class QuizXhr extends Control
 	public function editquest()
 	{
 		if ($this->session->mayEditQuiz()) {
-			if ($quest = $this->model->getQuestion($_GET['id'])) {
+			if ($quest = $this->quizGateway->getQuestion($_GET['id'])) {
 				$this->dataHelper->setEditData($quest);
 				$dia = new XhrDialog();
 
@@ -268,7 +267,7 @@ class QuizXhr extends Control
 	public function abort()
 	{
 		if ($this->session->may()) {
-			$this->model->abortSession($_GET['sid']);
+			$this->quizSessionGateway->abortSession($_GET['sid'], $this->session->id());
 
 			return array(
 				'status' => 1,
@@ -313,17 +312,17 @@ class QuizXhr extends Control
 		/*
 		 * First we want to check if there is a quiz session that the user has lost?
 		 */
-		if ($session = $this->quizGateway->getExistingSession($_GET['qid'], $this->session->id())) {
+		if ($session = $this->quizSessionGateway->getRunningSession($_GET['qid'], $this->session->id())) {
 			// if yes, reinitiate the running quiz session
 			$this->session->set('quiz-id', (int)$_GET['qid']);
 			$this->session->set('quiz-questions', $session['quiz_questions']);
 			$this->session->set('quiz-index', $session['quiz_index']);
 			$this->session->set('quiz-session', $session['id']);
-			$easymode = false;
+			$easyMode = false;
 			if ($session['easymode'] == 1 && (int)$_GET['qid'] == 1) {
-				$easymode = true;
+				$easyMode = true;
 			}
-			$this->session->set('quiz-easymode', $easymode);
+			$this->session->set('quiz-easymode', $easyMode);
 
 			/*
 			 * Make a little output that the user can continue the quiz
@@ -344,7 +343,7 @@ class QuizXhr extends Control
 		/*
 		 * Otherwise, we start a new quiz session
 		 */
-		if ($quiz = $this->model->getQuiz($_GET['qid'])) {
+		if ($quiz = $this->quizGateway->getQuiz($_GET['qid'])) {
 			/*
 			 * if foodsaver quiz, user can choose between easy and quick mode
 			*/
@@ -413,11 +412,11 @@ class QuizXhr extends Control
 
 		$dia->addAbortButton();
 
-		if ($this->session->get('hastodoquiz-id') == 1) {
+		if ($this->session->get('hastodoquiz-id') == Role::FOODSAVER) {
 			$dia->addButton('Jetzt mit dem Quiz meine Rolle als Foodsaver bestätigen', 'goTo(\'/?page=settings&sub=upgrade/up_fs\');');
-		} elseif ($this->session->get('hastodoquiz-id') == 2) {
+		} elseif ($this->session->get('hastodoquiz-id') == Role::STORE_MANAGER) {
 			$dia->addButton('Jetzt mit dem Quiz meine Rolle als Betriebsverantwortliche*r bestätigen', 'goTo(\'/?page=settings&sub=upgrade/up_bip\');');
-		} elseif ($this->session->get('hastodoquiz-id') == 3) {
+		} elseif ($this->session->get('hastodoquiz-id') == Role::AMBASSADOR) {
 			$dia->addButton('Jetzt mit dem Quiz meine Rolle als Botschafter*In bestätigen', 'goTo(\'/?page=settings&sub=upgrade/up_bot\');');
 		}
 
@@ -431,19 +430,19 @@ class QuizXhr extends Control
 	public function quizpopup()
 	{
 		if ($this->session->may('fs')) {
-			$count = (int)$this->model->qOne('SELECT COUNT(id) FROM fs_quiz_session WHERE foodsaver_id = ' . (int)$this->session->id() . ' AND quiz_id = ' . (int)$this->session->get('hastodoquiz-id') . ' AND `status` = 1');
-			if ($count == 0) {
+			$nextRole = $this->session->get('hastodoquiz-id');
+			if (!$this->quizGateway->hasPassedQuiz($this->session->id(), $nextRole)) {
 				$dia = new XhrDialog();
 				$dia->addOpt('width', 720);
 				$content_id = 18;
 				$dia->addAbortButton();
 
-				if ($this->session->get('hastodoquiz-id') == 1) {
+				if ($nextRole == Role::FOODSAVER) {
 					$dia->addButton('Ja, ich möchte jetzt mit dem Quiz meine Rolle als Foodsaver bestätigen.', 'goTo(\'/?page=settings&sub=upgrade/up_fs\');');
-				} elseif ($this->session->get('hastodoquiz-id') == 2) {
+				} elseif ($nextRole == Role::STORE_MANAGER) {
 					$content_id = 34;
 					$dia->addButton('Ja, ich möchte jetzt mit dem Quiz meine Rolle als Betriebsverantwortliche/r bestätigen.', 'goTo(\'/?page=settings&sub=upgrade/up_bip\');');
-				} elseif ($this->session->get('hastodoquiz-id') == 3) {
+				} elseif ($nextRole == Role::AMBASSADOR) {
 					$content_id = 35;
 					$dia->addButton('Ja, ich möchte jetzt mit dem Quiz meine Rolle als Botschafter*In bestätigen.', 'goTo(\'/?page=settings&sub=upgrade/up_bot\');');
 				}
@@ -464,7 +463,7 @@ class QuizXhr extends Control
 	public function addcomment()
 	{
 		if ($this->session->may() && !empty($_GET['comment']) && (int)$_GET['id'] > 0) {
-			$this->model->addUserComment((int)$_GET['id'], $_GET['comment']);
+			$this->quizGateway->addUserComment((int)$_GET['id'], $this->session->id(), $_GET['comment']);
 
 			return array(
 				'status' => 1,
@@ -494,15 +493,15 @@ class QuizXhr extends Control
 			 * If the quiz index is 0 we have to start a new quiz session
 			 */
 
-			$easymode = 0;
+			$easyMode = 0;
 			if ($this->session->get('quiz-easymode')) {
-				$easymode = 1;
+				$easyMode = 1;
 			}
 
 			if ($i == 0) {
-				$quuizz = $this->model->getQuiz($this->session->get('quiz-id'));
+				$quuizz = $this->quizGateway->getQuiz($this->session->get('quiz-id'));
 				// init quiz session in DB
-				if ($id = $this->quizGateway->initQuizSession($this->session->id(), $this->session->get('quiz-id'), $quiz, $quuizz['maxfp'], $quuizz['questcount'], $easymode)) {
+				if ($id = $this->quizSessionGateway->initQuizSession($this->session->id(), $this->session->get('quiz-id'), $quiz, $quuizz['maxfp'], $quuizz['questcount'], $easyMode)) {
 					$this->session->set('quiz-session', $id);
 				}
 			}
@@ -558,7 +557,7 @@ class QuizXhr extends Control
 				$comment = strip_tags($_GET['commentanswers'] . $_GET['comment']);
 
 				// if yes lets store in the db
-				$this->model->addUserComment((int)$_GET['qid'], $comment);
+				$this->quizGateway->addUserComment((int)$_GET['qid'], $this->session->id(), $comment);
 			}
 
 			/*
@@ -567,13 +566,13 @@ class QuizXhr extends Control
 			if (isset($_GET['special'])) {
 				// make a break
 				if ($_GET['special'] == 'pause') {
-					$this->model->updateQuizSession($this->session->get('quiz-session'), $quiz, $i);
+					$this->quizSessionGateway->updateQuizSession($this->session->get('quiz-session'), $quiz, $i);
 
 					return $this->pause();
 				}
 
 				if ($_GET['special'] == 'result') {
-					$this->model->updateQuizSession($this->session->get('quiz-session'), $quiz, $i);
+					$this->quizSessionGateway->updateQuizSession($this->session->get('quiz-session'), $quiz, $i);
 
 					return $this->resultNew($quiz[($i - 1)], $dia->getId());
 				}
@@ -585,10 +584,10 @@ class QuizXhr extends Control
 			 */
 			if (isset($quiz[$i])) {
 				// get the question
-				if ($question = $this->model->getQuestion($quiz[$i]['id'])) {
+				if ($question = $this->quizGateway->getQuestion($quiz[$i]['id'])) {
 					// get possible answers
 					$comment_aswers = '';
-					if ($answers = $this->model->getAnswers($question['id'])) {
+					if ($answers = $this->quizGateway->getAnswers($question['id'])) {
 						// random sorting for the answers
 						shuffle($answers);
 
@@ -606,7 +605,7 @@ class QuizXhr extends Control
 
 						// update quiz session
 						$session_id = $this->session->get('quiz-session');
-						$this->model->updateQuizSession($session_id, $quiz, $i);
+						$this->quizSessionGateway->updateQuizSession($session_id, $quiz, $i);
 						$this->session->set('quiz-quest-start', time());
 
 						/*
@@ -665,7 +664,7 @@ class QuizXhr extends Control
 							var counter = null;
 						';
 
-						if ($easymode == 0) {
+						if ($easyMode == 0) {
 							$quizbreath = '
 							$(\'#quizwrapper\').hide();
 							$(\'#quizbreath\').show();
@@ -885,16 +884,16 @@ class QuizXhr extends Control
 			return false;
 		}
 
-		if ($quiz = $this->model->getQuiz($this->session->get('quiz-id'))) {
+		if ($quiz = $this->quizGateway->getQuiz($this->session->get('quiz-id'))) {
 			if ($questions = $this->session->get('quiz-questions')) {
-				if ($rightQuestions = $this->model->getRightQuestions($this->session->get('quiz-id'))) {
+				if ($rightQuestions = $this->quizGateway->getRightQuestions($this->session->get('quiz-id'))) {
 					$explains = array();
-					$fp = 0;
+					$failurePoints = 0;
 					$question_number = 0;
 					foreach ($questions as $q_key => $q) {
 						++$question_number;
 						$valid = $this->validateAnswer($rightQuestions, $q);
-						$fp += $valid['fp'];
+						$failurePoints += $valid['fp'];
 						foreach ($valid['explain'] as $e) {
 							if (!isset($explains[$q['id']])) {
 								$explains[$q['id']] = $rightQuestions[$q['id']];
@@ -908,7 +907,7 @@ class QuizXhr extends Control
 					}
 				}
 
-				$this->model->finishQuiz($this->session->get('quiz-session'), $questions, $explains, $fp, $quiz['maxfp']);
+				$this->quizSessionGateway->finishQuizSession($this->session->get('quiz-session'), $questions, $explains, $failurePoints, $quiz['maxfp']);
 
 				return array(
 					'status' => 1,
@@ -929,9 +928,9 @@ class QuizXhr extends Control
 			}
 		}
 		// get the question
-		if ($quest = $this->model->getQuestion($question['id'])) {
+		if ($quest = $this->quizGateway->getQuestion($question['id'])) {
 			// get possible answers
-			if ($answers = $this->model->getAnswers($question['id'])) {
+			if ($answers = $this->quizGateway->getAnswers($question['id'])) {
 				$joke = false;
 				if ($question['fp'] == 0) {
 					$joke = true;
@@ -1090,7 +1089,7 @@ class QuizXhr extends Control
 		// wie viel Prozent sind falsch?
 		$percent = $this->percentFrom($checkCount, $wrongAnswers);
 
-		$fp = $this->percentTo($question['fp'], $percent);
+		$failurePoints = $this->percentTo($question['fp'], $percent);
 
 		// wenn alles falsch angeklickt wurde, das aber nicht stimmt, gibt's die volle Fehlerpunktezahl
 		if (
@@ -1098,18 +1097,18 @@ class QuizXhr extends Control
 			||
 			(!$everything_false && (int)$question['noco'] > 0)
 		) {
-			$fp = $question['fp'];
+			$failurePoints = $question['fp'];
 			$percent = 100;
 		}
 
 		// fix alle Fragen sind neutral
 		if ($allNeutral) {
-			$fp = 0;
+			$failurePoints = 0;
 			$percent = '0';
 		}
 
 		return array(
-			'fp' => $fp,
+			'fp' => $failurePoints,
 			'explain' => $explains,
 			'percent' => $percent
 		);
@@ -1119,7 +1118,7 @@ class QuizXhr extends Control
 	{
 		$count_questions = $count;
 
-		if ($questions = $this->model->getQuestionMetas($quiz_id)) {
+		if ($questions = $this->quizGateway->getQuestionMetas($quiz_id)) {
 			// Wie viele Fragen gibt es insgesamt?
 			$summe = 0;
 			foreach ($questions['meta'] as $key => $m) {
@@ -1127,13 +1126,13 @@ class QuizXhr extends Control
 			}
 
 			$out = array();
-			// Prozentanzeil von jeder Fragenart
+			// Prozentanteil von jeder Fragenart
 			foreach ($questions['meta'] as $key => $m) {
 				$percent = round($this->percentFrom($summe, $m));
 
 				$count = round($this->percentTo($count_questions, $percent));
 
-				if ($rquest = $this->model->getRandomQuestions($count, $key, $quiz_id)) {
+				if ($rquest = $this->quizGateway->getRandomQuestions($count, $key, $quiz_id)) {
 					foreach ($rquest as $r) {
 						$out[] = $r;
 					}
@@ -1173,13 +1172,13 @@ class QuizXhr extends Control
 				 [fp] => 3
 			 */
 			if (isset($_GET['text'], $_GET['fp'], $_GET['id'])) {
-				$fp = (int)$_GET['fp'];
+				$failurePoints = (int)$_GET['fp'];
 				$text = strip_tags($_GET['text']);
 				$duration = (int)$_GET['duration'];
-				$wikilink = strip_tags($_GET['wikilink']);
+				$wikiLink = strip_tags($_GET['wikilink']);
 
 				if (!empty($text)) {
-					$this->model->updateQuestion($_GET['id'], $_GET['qid'], $text, $fp, $duration, $wikilink);
+					$this->quizGateway->updateQuestion($_GET['id'], $_GET['qid'], $text, $failurePoints, $duration, $wikiLink);
 					$this->flashMessageHelper->info('Frage wurde geändert');
 
 					return array(