+++ CHANGELOG.md
@@ -79,6 +79,7 @@
 - Converted nightly maintenance methods deactivateOldBaskets() and deleteUnconfirmedFetchDates() into gateway !976 @jofranz
 - Refactored team page. Got rid of legacy methods !974 @jofranz
 - Refactored fetch weight menu handling and moved weight methods into a helper class !1002 @jofranz
+- Refactored profile from QuizModel to QuizGateway !998 #9 @svenpascal
 
 ## Dev/Test/CI stuff
 - enable functional tests (symfony kernel running inside conception; for limits see inside tests/functional folder) !884 @NerdyProjects